fix compiler warnings
MozReview-Commit-ID: 43E32Sbt4US

+++ intl/hyphenation/hyphen/hyphen.c
@@ -329,7 +329,7 @@ void hnj_hyphen_load_line(char * buf, HyphenDict * dict, HashTab * hashtab) {
                 int pu = -1;        /* unicode character position */
                 int ps = -1;        /* unicode start position (original replindex) */
                 int pc = (*word == '.') ? 1: 0; /* 8-bit character position */
-                for (; pc < (strlen(word) + 1); pc++) {
+                for (; pc < (int)(strlen(word) + 1); pc++) {
                 /* beginning of an UTF-8 character (not '10' start bits) */
                     if ((((unsigned char) word[pc]) >> 6) != 2) pu++;
                     if ((ps < 0) && (replindex == pu)) {
